pyproject.toml
[tool.poetry]
name = "TalkBot"
version = "0.1.0"
description = "Hi,TalkBot~"
authors = ["RedElectricity <redelectricity@unaltered.xyz>"]

[tool.poetry.dependencies]
python = "^3.9"
fuzzywuzzy = "^0.18.0"
jieba = "^0.42.1"
httpx = "^0.18.2"
PyYAML = "^5.4.1"
tensorflow = "^2.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
